chore: API key types
